refactor: Introduce mobilityprofile module
