+++ i18n/mobilityprofile/en-US.yml
@@ -1,0 +1,64 @@
+_id: en-US
+_name: English
+
+actions:
+  user:
+    getDependentUserInfoFailed: Error getting mobility profile information.
+components:
+  CompanionsPane:
+    addNewCompanion: Add a new travel companion
+    companionAlreadyAdded: You already have a companion with email {email}
+    companionExplanation: >
+      Invite an existing G-MAP user to be a travel companion by entering their
+      email. When they accept, their status will change to "verified", and you
+      can share your trip status and plan trips based on one another's mobility
+      profile.
+    confirmDeleteCompanion: Do you want to delete companion {email}?
+    currentCompanionsLabel: "Current travel companions:"
+    deleteCompanion: Delete {email}
+    noCompanions: You do not have any existing travel companions.
+    submitNewCompanion: Send invitation
+    title: Travel companions
+  MobilityProfile:
+    DevicesPane:
+      devices:
+        cane: Cane
+        crutches: Crutches
+        electric wheelchair: Electric wheelchair
+        manual walker: Manual walker
+        manual wheelchair: Manual/traditional wheelchair
+        mobility scooter: Mobility scooter
+        none: No assistive device
+        service animal: Service animal
+        stroller: Stroller
+        wheeled walker: Wheeled walker
+        white cane: White cane
+      prompt: Do you regularly use a mobility assistive device? (Check all that apply)
+    LimitationsPane:
+      mobilityPrompt: >-
+        Do you have any mobility limitations that cause you to walk more slowly
+        or more carefully than other people?
+      visionLimitations:
+        legally-blind: Legally blind
+        low-vision: Low-vision
+        none: None
+      visionPrompt: Do you have any vision limitations?
+    MobilityPane:
+      button: Edit your mobility profile
+      header: Mobility Profile
+      mobilityDevices: "Mobility devices: "
+      mobilityLimitations: "Mobility limitations: "
+      planTripDescription: >-
+        If you have a travel companion, you can choose to plan this trip
+        according to their mobility profile.
+      visionLimitations: "Vision limitations: "
+    dropdownLabel: "User mobility profile:"
+    intro: >-
+      Please answer a few questions to customize the trip planning experience to
+      your needs and preferences.
+    myself: Myself
+    title: Define Your Mobility Profile
+  TripCompanionsPane:
+    companionLabel: "Companion on this trip:"
+    observersLabel: "Observers watching this trip:"
+    primaryLabel: "Primary traveler: "
