+++ lib/components/form/advanced-settings-mobility-profile.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+import AppModule from '../app/app-module'
+import DependentSelector from '../user/mobility-profile/dependent-selector'
+
+const AdvancedSettingsMobilityProfile = (): JSX.Element | null => (
+  <AppModule name="mobilityprofile">
+    <DependentSelector />
+  </AppModule>
+)
+
+export default AdvancedSettingsMobilityProfile
