+++ lib/util/i18n-loader.js
@@ -7,13 +7,20 @@ const i18nModules = import.meta.glob('../../i18n/*.yml', { import: 'default' })
 const i18nAccountModules = import.meta.glob('../../i18n/account/*.yml', {
   import: 'default'
 })
+const i18nMobilityModules = import.meta.glob(
+  '../../i18n/mobilityprofile/*.yml',
+  {
+    import: 'default'
+  }
+)
 const i18nOtpUiModules = import.meta.glob(
   '../../node_modules/@opentripplanner/**/i18n/*.yml',
   { import: 'default' }
 )
 
 const appModules = {
-  account: i18nAccountModules
+  account: i18nAccountModules,
+  mobilityprofile: i18nMobilityModules
 }
 
 /**
